+++ CppLibrary/include/aiRunner.h
@@ -6,8 +6,8 @@
 #include <vector>
 
 extern "C" bool InitModel(const char* modelPath, bool cpu_use);
-extern "C" void RunModelInt(int* data, size_t num_elements);
-extern "C" void RunModelFloat(float* data, size_t num_elements);
-extern "C" void RunModelDouble(double* data, size_t num_elements);
+extern "C" void RunModelInt(int* data, int num_elements);
+extern "C" void RunModelFloat(float* data, int  num_elements);
+extern "C" void RunModelDouble(double* data, int num_elements);
 extern "C" const float* GetFlattenedOutput(int* size);
 extern "C" const int64_t* GetOriginalShape(int* size);
